Rename packages without client suffix
