+++ azure/hostedservice/entities.go
@@ -1,4 +1,4 @@
-package hostedserviceclient
+package hostedservice
 
 import (
 	"encoding/xml"
