+++ azure/storage/util.go
@@ -1,4 +1,4 @@
-package storageclient
+package storage
 
 import (
 	"crypto/hmac"
