+++ azure/storage/util_test.go
@@ -1,4 +1,4 @@
-package storageclient
+package storage
 
 import (
 	"io/ioutil"
