+++ azure/storageservice/client.go
@@ -1,4 +1,4 @@
-package storageserviceclient
+package storageservice
 
 import (
 	"encoding/base64"
